partly Revert "split timeline widget"
This reverts commit 2950f60b9a20f9aebd7ae55b3c4910cd15d0cc12.

+++ src/core/engine.h
@@ -62,8 +62,8 @@ class Engine
         double get_bpm(){return m_bpm;};
         int m_length;
         int get_length(){return m_length;};
-        int m_queue_length;
-
+        double m_queue_length;
+        
         int m_n_loops;
         std::list <Loop> m_loops;
